Adding FastCGI library to enable exposing parts of the functionality and control of the DisplayCluster as web services.
In order to expose the display cluster as a service, a web server with
FastCGI capabilities (Apache, NGINX), must be setup in front of the cluster.
This web server will handle static content requests, and will forward
service calls to the Display cluster using the FastCGI protocol

The FastCGI protocol allow use to directly expose functionality written
in C++ as a web service.

Change-Id: Ifaf0bca249cc3aa689aa5c143635d882cbba1252